refactor: move security handlers to their own package
